refactor: 인덱스 추가 및 재고 집계 수량 쿼리 개선

+++ hub/src/main/java/com/klp/hub/inventory/domain/repository/dto/InventoryAvailability.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 public record InventoryAvailability(
     UUID productId,
     UUID hubId,
-    Integer availableQuantity
+    Long availableQuantity
 ) {
 
 }
